Refactor Favorite endpoint tests into single file
